Redigital Image Pipeline/requirements.txt
Pillow==10.2.0
PyPDF2==3.0.1
PyMuPDF==1.23.8 